Ideal Angle Vector - CI
To understand customer's preferences within a category, we as an analyst for Car Dealership would like to figure out what attributes actually matter to the customer. Once we understand the important attributes, we would know the direction in which to take the company in, so that we increase customer preference for us.

The idea is to first understand the weightage given to each attribute. For ex: Amongst 'Attractive',	'Quiet','Unreliable','Poorly Built','Interesting' etc, what are the component scores? Then run the preference regression to estimate how benefits drive overall preferences.

Prinicipal Component Analysis
As a first step, we create the correlation matrix which we will decompose into eigenvalues. Then chose the eigenvalues that are greater than 1. These are the values that will actually be significant for the preferences. We can also confirm this using a scree plot. anything above the elbow is retained.
With the retained factors, regression for preferences is run to check which eigen vectors turn out significant.

With the significant vectors, the brands map is built.
Brand maps inform which brands compete strongly with others. Proximity means competition. So now the goal is get a sense of direction of improvement. This directional angle is called ideal vector angle. It is to make sure while customer preference increases the competition doesn't. It's a good way to understand the trade offs a customer might be willing to make.