Webscraping Ebay using python and sql
The goal of this project was to scrape product data from ebay. and then analyze the stats summary to determine whether list price's visibility has any correlation with the seller score, items sold and price of the items.
Keeping the condition of the item constant, if the list price is provided by the seller then, the seller score, items sold and item price will be higher as compared to when list price is not provided.