Create Predicting Loan Propensity
Goal:
To analyze how well we can predict the propensity to get a loan from the data collected. Are
well-connected households more likely to seek a loan (they are comfortable with external
finance)? Or are they less likely (their network provides the support they need already)?


Since this analysis is on observational data it's hard to separate out the Treatment Effect. The solution used here is Double Lasso. The first Lasso captures the variation of X in d. this is stored in the residuals. 
In the second lasso, the residuals, d - treatment effect and X are all regressed again to get the pure treatment effect. The double lasso will drop a few variables but we don't have to worry about those because their effect has already been captured in the residuals from the first Lasso.

Transform the degree to create the treatment variable d.
Model Building – Build a model to predict the treatment variable d from the control.
Interact all control variables with one another up to third degree (~.^3).
Prediction – Using predictions from above in an estimator for the effect of the treatment
variable on loan with visualizations and explanations.
 Model Comparison – Comparing the results of part-3 from different models ( lasso +
CV vs naïve lasso) for the loan on treatment variable and the dependent variables.
Explain why they are similar or different.
Bootstrap – Bootstrap  estimator